+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Don't mark master instance as unhealthy if local etcd instance is unresponsive but the whole etcd cluster is down. 
+- Don't mark master instance as unhealthy if local etcd instance is unresponsive but the whole etcd cluster is also down.
+- Don't mark master instance as unhealthy if local API server instance is unresponsive but the whole API server is also down.
 
 ## [14.8.0] - 2023-03-02